+++ CHANGELOG.md
@@ -5,6 +5,12 @@ Headlines: Added, Changed, Deprecated, Removed, Fixed, Security
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.5.1] - 2023-08-09
+
+### Fixed
+
+- In rare cases, timers weren't resetted when redirecting to the same route.
+
 ## [1.5.0] - 2023-07-23
 
 ### Fixed